+++ chart/tests/test_git_sync_worker.py
