+++ chart/tests/test_scheduler.py
